Rename hpx_main.hpp to wrap_main.hpp
