+++ wrap/src/main.cpp
@@ -15,7 +15,7 @@
 
 // when statically linking the default definition of main happens on demand only
 #if !defined(HPX_HAVE_STATIC_LINKING)
-#include <hpx/hpx_main_impl.hpp>
+#include <hpx/wrap_main_impl.hpp>
 #endif
 
 #endif
